+++ pombola/experiments/views.py
@@ -1,5 +1,11 @@
 import json
+from random import randint
 import re
+import sys
+
+from django.core.urlresolvers import reverse
+from django.utils.http import urlquote
+from django.views.generic.base import RedirectView
 
 from pombola.experiments.models import Experiment
 from pombola.feedback.models import Feedback
@@ -125,6 +131,49 @@ class ExperimentFormSubmissionMixin(ExperimentViewDataMixin):
                            'label': self.form_key})
         return super(ExperimentFormSubmissionMixin,
                      self).form_valid(form)
+
+class ExperimentShare(ExperimentViewDataMixin, RedirectView):
+    """For recording & enacting Facebook / Twitter share actions"""
+
+    permanent = False
+
+    def get_redirect_url(self, *args, **kwargs):
+        social_network = sanitize_parameter(
+            key='n',
+            parameters=self.request.GET,
+            allowed_values=('facebook', 'twitter'))
+        share_key = "{0:x}".format(randint(0, sys.maxint))
+        self.create_event({'category': 'share-click',
+                           'action': 'click',
+                           'label': social_network,
+                           'share_key': share_key})
+        path = reverse(self.base_view_name)
+        built = self.request.build_absolute_uri(path)
+        built += '?via=' + share_key
+        url_parameter = urlquote(built, safe='')
+        url_formats = {
+            'facebook': "https://www.facebook.com/sharer/sharer.php?u={0}",
+            'twitter': "http://twitter.com/share?url={0}"}
+        return url_formats[social_network].format(url_parameter)
+
+
+class ExperimentSurvey(ExperimentViewDataMixin, RedirectView):
+    """For redirecting to the Qualtrics survey"""
+
+    permanent = False
+
+    def get_redirect_url(self, *args, **kwargs):
+        self.create_event({'category': 'take-survey',
+                           'action': 'click',
+                           'label': 'take-survey'})
+        prefix = self.session_key_prefix
+        sid = self.qualtrics_sid
+        url = "http://survey.az1.qualtrics.com/SE/?SID={0}&".format(sid)
+        url += "&".join(
+            k + "=" + self.request.session.get(prefix + ':' + k, '?')
+            for k in ['user_key', 'variant'] + self.demographic_keys.keys()
+        )
+        return url
 
 
 def sanitize_parameter(key, parameters, allowed_values, default_value=None):
